Change oauth_file path to a more portable one
The previous one only worked if you started mopidy as a normal user. If
it was started as a service, it would try to write to a non-existent
directory.

The OAuth token is also moved to the cache directory instead as that is
arguably the right place for it to reside.